chore(Queries): query result component refactoring [YTFRONT-4423]
