+++ packages/ui/src/ui/pages/query-tracker/QueryResultsVisualization/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import withLazyLoading from '../../../hocs/withLazyLoading';
+import type {QueryItem} from '../module/api';
 
 const QueryResultsVisualizationLazy = withLazyLoading(
     React.lazy(async () => {
@@ -13,7 +14,7 @@ const QueryResultsVisualizationLazy = withLazyLoading(
     }),
 );
 
-export const CUSTOM_QUERY_REQULT_TAB = {
+export const QUERY_RESULT_CHART_TAB = {
     title: 'Chart',
-    renderContent: (props: any) => <QueryResultsVisualizationLazy {...props} />,
+    renderContent: (params: {query: QueryItem}) => <QueryResultsVisualizationLazy {...params} />,
 };
